Fix Vehicle Analysis Method Calls
