+++ RetireSimple.Engine/Data/Base/InvestmentVehicle.cs
@@ -58,9 +58,6 @@ namespace RetireSimple.Engine.Data.Base {
 			//TODO allow contributing to other investments as well, requires post-processing logic
 			var cashSim = SimulateCashContributions(combinedOptions);
 
-			//TODO Pre-process expenses/transfers (from investment in vehicle) to adjust for taxes
-			// This will be more relevant when expenses become a factor in analysis.
-
 			var preTaxModel = GeneratePreTaxModels(combinedOptions, containedModels, cashSim);
 			var postTaxModel = GeneratePostTaxModels(combinedOptions, containedModels, cashSim);
