+++ RetireSimple.Engine/Data/InvestmentVehicle/Vehicle403b.cs
@@ -7,12 +7,13 @@ namespace RetireSimple.Engine.Data.InvestmentVehicle {
 		public override InvestmentModel GeneratePostTaxModels(OptionsDict options,
 														List<InvestmentModel> models,
 														List<decimal>? cashContribution = null)
-		=> VehicleDefaultAS.GeneratePreTaxModelDefault(options, models, cashContribution);
+			=> VehicleDefaultAS.GeneratePostTaxModelDefault(options, models, cashContribution);
 		public override InvestmentModel GeneratePreTaxModels(OptionsDict options,
 															List<InvestmentModel> models,
 															List<decimal>? cashContribution = null)
-			=> VehicleDefaultAS.GeneratePostTaxModelDefault(options, models, cashContribution);
+			=> VehicleDefaultAS.GeneratePreTaxModelDefault(options, models, cashContribution);
+
 		public override List<decimal> SimulateCashContributions(OptionsDict options)
-			=> VehicleDefaultAS.SimulateCashContributionsDefault(this, options);
+			=> VehicleDefaultAS.SimulateCashContributions401kLike(this, options);
 	}
 }
